# schemas/openapi-v3.yaml
components:
  schemas:
    business.RenameManagedFileRequest:
      properties:
        currents:
          items:
            type: string
          type: array
          uniqueItems: false
        koji:
          $ref: '#/components/schemas/models.Koji'
      type: object
    handlers.BulkConversionRequest:
      properties:
        items:
          description: 変換対象のアイテム
          items:
            $ref: '#/components/schemas/handlers.IDGenerationRequest'
          type: array
          uniqueItems: false
      type: object
    handlers.BulkConversionResponse:
      properties:
        errorCount:
          description: エラー数
          type: integer
        processedAt:
          description: 処理日時
          type: string
        results:
          description: 変換結果
          items:
            $ref: '#/components/schemas/handlers.IDGenerationResponse'
          type: array
          uniqueItems: false
        successCount:
          description: 成功数
          type: integer
        totalCount:
          description: 総数
          type: integer
      type: object
    handlers.IDGenerationRequest:
      properties:
        companyName:
          description: 会社名
          type: string
        fullPath:
          description: フルパス（パスID生成用）
          type: string
        locationName:
          description: 場所名
          type: string
        startDate:
          description: ISO 8601形式
          type: string
      type: object
    handlers.IDGenerationResponse:
      properties:
        created:
          description: 生成日時
          type: string
        errorMsg:
          description: エラーメッセージ
          type: string
        id:
          description: 生成されたID
          type: string
        idType:
          description: IDの種類（koji/path）
          type: string
        isValid:
          description: 有効性
          type: boolean
        length:
          description: ID長
          type: integer
        method:
          description: 生成方法
          type: string
        source:
          description: 生成元データ
          type: string
      type: object
    handlers.IDValidationRequest:
      properties:
        companyName:
          description: 会社名
          type: string
        fullPath:
          description: フルパス
          type: string
        id:
          description: 検証対象のID
          type: string
        locationName:
          description: 場所名
          type: string
        startDate:
          description: ISO 8601形式
          type: string
      type: object
    handlers.IDValidationResponse:
      properties:
        comparedAt:
          description: 比較日時
          type: string
        errorMsg:
          description: エラーメッセージ
          type: string
        expectedId:
          description: 期待されるID
          type: string
        isValid:
          description: 有効性
          type: boolean
        method:
          description: 検証方法
          type: string
        needsSync:
          description: 同期が必要かどうか
          type: boolean
        providedId:
          description: 提供されたID
          type: string
      type: object
    models.AssistFile:
      properties:
        current:
          description: 現在のファイル名
          example: 工事.xlsx
          type: string
        currentPath:
          description: FileService.BasePathからの相対パス
          example: 豊田築炉/2-工事/2025-0618 豊田築炉 名和工場
          type: string
        desired:
          description: 希望するファイル名
          example: 2025-0618 豊田築炉 名和工場.xlsx
          type: string
        desiredPath:
          description: FileService.BasePathからの相対パス
          example: 豊田築炉/2-工事/2025-0618 豊田築炉 名和工場
          type: string
      type: object
    models.Company:
      description: 工事会社の基本情報をファイル名から取得したモデル
      properties:
        address:
          example: 愛知県名古屋市熱田区三本松町1-1
          type: string
        assists:
          description: 補助ファイルフィールド
          items:
            $ref: '#/components/schemas/models.AssistFile'
          type: array
          uniqueItems: false
        category:
          $ref: '#/components/schemas/models.CompanyCategory'
        email:
          example: info@toyotachikuro.jp
          type: string
        folderName:
          description: Identity fields. FolderName is linked ShortName and Category.
          example: 0 豊田築炉
          type: string
        id:
          description: ID is generated from FolderName.
          example: TC001
          type: string
        legalName:
          description: Database file fields
          example: 有限会社 豊田築炉
          type: string
        phone:
          example: 052-681-8111
          type: string
        postalCode:
          example: 456-0001
          type: string
        shortName:
          example: 豊田築炉
          type: string
        tags:
          example:
          - '[''元請け'''
          - ' ''製造業'']'
          items:
            type: string
          type: array
          uniqueItems: false
        website:
          example: https://www.toyotachikuro.jp
          type: string
      type: object
    models.CompanyCategory:
      example: 4
      type: integer
      x-enum-comments:
        CompanyCategoryAgency: '1: 応援会社'
        CompanyCategoryLease: '5: リース会社'
        CompanyCategoryOther: '9: その他'
        CompanyCategoryPeer: '2: 同業者（築炉会社）'
        CompanyCategoryPersonal: '3: 一人親方'
        CompanyCategoryPrime: '4: 元請け'
        CompanyCategoryRecruit: '8: 求人会社'
        CompanyCategorySales: '6: 販売会社'
        CompanyCategorySales2: '7: 販売会社（重複）'
        CompanyCategorySpecial: '0: 特別(自社・組合・未定)'
      x-enum-varnames:
      - CompanyCategorySpecial
      - CompanyCategoryAgency
      - CompanyCategoryPeer
      - CompanyCategoryPersonal
      - CompanyCategoryPrime
      - CompanyCategoryLease
      - CompanyCategorySales
      - CompanyCategorySales2
      - CompanyCategoryRecruit
      - CompanyCategoryOther
    models.FileInfo:
      description: ファイルまたはディレクトリの情報
      properties:
        isDirectory:
          description: Whether this item is a directory
          example: true
          type: boolean
        modifiedTime:
          $ref: '#/components/schemas/models.Timestamp'
        name:
          description: Name of the file or folder
          example: 2025-0618 豊田築炉 名和工場
          type: string
        path:
          description: FileService.BasePathからの相対パス
          example: 豊田築炉/2-工事/2025-0618 豊田築炉 名和工場
          type: string
        size:
          description: Size of the file in bytes
          example: 4096
          type: integer
      type: object
    models.Koji:
      description: 拡張属性を持つ工事プロジェクトフォルダー情報
      properties:
        assists:
          items:
            $ref: '#/components/schemas/models.AssistFile'
          type: array
          uniqueItems: false
        companyName:
          example: 豊田築炉
          type: string
        description:
          example: 工事関連の資料とドキュメント
          type: string
        endDate:
          $ref: '#/components/schemas/models.Timestamp'
        folderName:
          description: Identity fields. FolderName is linked StartDate, CompanyName
            and LocationName.
          example: 豊田築炉/2-工事/2025-0618 豊田築炉 名和工場
          type: string
        id:
          description: ID, Status is calculated from FolderName.
          example: TC618
          type: string
        locationName:
          example: 名和工場
          type: string
        startDate:
          $ref: '#/components/schemas/models.Timestamp'
        status:
          example: 進行中
          type: string
        tags:
          example:
          - '[''工事'''
          - ' ''豊田築炉'''
          - ' ''名和工場'']'
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    models.Timestamp:
      description: Database file fields
      properties:
        time.Time:
          type: string
      type: object
externalDocs:
  description: ""
  url: ""
info:
  description: ファイルエントリの管理と閲覧のためのAPI
  title: Penguin ファイルシステム管理API
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/id-sync/bulk-convert:
    post:
      description: 複数のアイテムを一括でID変換します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.BulkConversionRequest'
        description: 一括変換リクエスト
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.BulkConversionResponse'
          description: 正常なレスポンス
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: リクエストエラー
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: 一括ID変換
      tags:
      - ID同期
  /api/id-sync/config:
    get:
      description: ID同期システムの設定情報を取得します
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: 同期設定情報
      summary: 同期設定の取得
      tags:
      - ID同期
  /api/id-sync/generate-koji:
    post:
      description: 開始日・会社名・場所名から工事IDを生成します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.IDGenerationRequest'
        description: ID生成リクエスト
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.IDGenerationResponse'
          description: 正常なレスポンス
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: リクエストエラー
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: 工事IDの生成
      tags:
      - ID同期
  /api/id-sync/generate-path:
    post:
      description: フルパスからLen7 IDを生成します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.IDGenerationRequest'
        description: ID生成リクエスト
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.IDGenerationResponse'
          description: 正常なレスポンス
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: リクエストエラー
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: パスIDの生成
      tags:
      - ID同期
  /api/id-sync/mapping:
    get:
      description: フルパスIDとLen7 IDのマッピングテーブルを取得します
      parameters:
      - description: カンマ区切りのパスリスト
        in: query
        name: paths
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: IDマッピングテーブル
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: リクエストエラー
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: IDマッピングテーブルの取得
      tags:
      - ID同期
  /api/id-sync/validate:
    post:
      description: 提供されたIDが正しいかどうかを検証します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.IDValidationRequest'
        description: ID検証リクエスト
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.IDValidationResponse'
          description: 正常なレスポンス
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: リクエストエラー
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: IDの検証
      tags:
      - ID同期
  /business/attribute-filename:
    get:
      description: 属性ファイル名を取得します
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: 属性ファイル名
      summary: 属性ファイル名の取得
      tags:
      - ビジネス管理
  /business/base-path:
    get:
      description: ビジネスベースパスを取得します
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: ビジネスベースパス
      summary: ビジネスベースパスの取得
      tags:
      - ビジネス管理
  /business/categories:
    get:
      description: カテゴリー一覧を取得します
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  additionalProperties:
                    type: string
                  type: object
                type: array
          description: 正常なレスポンス
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: カテゴリー一覧の取得
      tags:
      - 会社管理
  /business/companies:
    get:
      description: 会社フォルダーの一覧を取得します
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.Company'
                type: array
          description: 正常なレスポンス
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: 会社一覧の取得
      tags:
      - 会社管理
    put:
      description: 会社情報を更新し、必要に応じてフォルダー名を変更します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.Company'
        description: 会社更新リクエスト
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Company'
          description: 更新後の会社データ
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: 会社情報の更新
      tags:
      - 会社管理
  /business/companies/{id}:
    get:
      description: 指定されたIDの会社詳細を取得します
      parameters:
      - description: 会社ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Company'
          description: 正常なレスポンス
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: 会社が見つからない
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: 会社詳細の取得
      tags:
      - 会社管理
  /business/files:
    get:
      description: 指定されたパスからファイルとフォルダーの一覧を取得します
      parameters:
      - description: 取得するディレクトリのパス
        in: query
        name: path
        schema:
          default: 2 工事
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.FileInfo'
                type: array
          description: 正常なレスポンス
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: ファイルエントリ一覧の取得
      tags:
      - ファイル管理
  /business/kojies:
    get:
      description: 工事一覧を取得します。filter=recentで最近の工事のみを取得できます。
      parameters:
      - description: 'フィルター (recent: 最近の工事のみ)'
        in: query
        name: filter
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.Koji'
                type: array
          description: 工事一覧
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: 工事一覧の取得
      tags:
      - 工事管理
    put:
      description: 更新対象フォルダー（Koji.FileInfo）のファイル情報をKojiのメンバ変数で更新します。
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.Koji'
        description: 工事データ
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Koji'
          description: 更新後の工事データ
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: 更新対象フォルダー（Koji.FileInfo）のファイル情報をKojiのメンバ変数で更新
      tags:
      - 工事管理
  /business/kojies/assist-files:
    put:
      description: AssistFileの名前を変更します。
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/business.RenameManagedFileRequest'
        description: 工事データと管理ファイル
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Koji'
          description: 更新後の工事データ
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: AssistFileの名前を変更
      tags:
      - 工事管理
  /kojies/{path}:
    get:
      description: Kojiを取得します。
      parameters:
      - description: Kojiフォルダーのファイル名
        in: path
        name: path
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Koji'
          description: Koji
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: サーバーエラー
      summary: Kojiの取得
      tags:
      - 工事管理
servers:
- url: http://localhost:8080/api